Allows custom keycodes in layer taps
This isn't really an implementation of layer taps with custom keycodes,
but a reasonable way to work around it. It still requires the user to specify
the layer tap keycode for e.g. capitalize in the `is_capitalize_key` function.
The difference is that the held key will go ignored.