Adding on a column again
Needs the Swedish letters a lot in the new job, which is difficult to
fit without using the sixth column. This goes back to using the pinkies
for some common keys, let's see how it feels.
Also puts shift back on a thumb. Space is moved to the right thumb, it
makes more sense with the setup of nav layer and using shift to get
extra symbols from the sym layer.

The nav layer should be refined more?